Moved all section components to modules folder
