+++ src/app/page.js
@@ -1,9 +1,9 @@
-import Contacts from '@/components/Contacts/Contacts';
-import Features from '@/components/Features/Features';
-import Gallery from '@/components/Gallery/Gallery';
-import Hero from '@/components/Hero/Hero';
-import Price from '@/components/Price/Price';
-import Reviews from '@/components/Reviews/Reviews';
+import Contacts from '@/modules/Contacts';
+import Features from '@/modules/Features';
+import Gallery from '@/modules/Gallery';
+import Hero from '@/modules/Hero';
+import Price from '@/modules/Price';
+import Reviews from '@/modules/Reviews';
 
 export default function Home() {
   return (
