--- src/components/Contacts/Contacts.jsx
@@ -1,4 +1,0 @@
-const Contacts = () => {
-  return <div>Contacts</div>;
-};
-export default Contacts;
