--- src/components/Footer/Footer.jsx
@@ -1,8 +1,0 @@
-const Footer = () => {
-  return (
-    <footer className='mx-auto max-w-[1440px]'>
-      <div className='container'>Footer</div>
-    </footer>
-  );
-};
-export default Footer;
