--- src/components/Gallery/Gallery.jsx
@@ -1,4 +1,0 @@
-const Gallery = () => {
-  return <div>Gallery</div>;
-};
-export default Gallery;
