--- src/components/Price/Price.jsx
@@ -1,4 +1,0 @@
-const Price = () => {
-  return <div>Price</div>;
-};
-export default Price;
