--- src/components/Reviews/Reviews.jsx
@@ -1,4 +1,0 @@
-const Reviews = () => {
-  return <div>Reviews</div>;
-};
-export default Reviews;
