+++ src/modules/Contacts/Contacts.jsx
@@ -1,0 +1,5 @@
+const Contacts = () => {
+  return <div>Contacts</div>;
+};
+
+export default Contacts;
