+++ src/modules/Contacts/index.js
@@ -1,0 +1,1 @@
+export { default } from './Contacts.jsx';
