+++ src/modules/Footer/Footer.jsx
@@ -1,0 +1,9 @@
+const Footer = () => {
+  return (
+    <footer className='mx-auto max-w-[1440px]'>
+      <div className='container'>Footer</div>
+    </footer>
+  );
+};
+
+export default Footer;
