+++ src/modules/Gallery/Gallery.jsx
@@ -1,0 +1,5 @@
+const Gallery = () => {
+  return <div>Gallery</div>;
+};
+
+export default Gallery;
