+++ src/modules/Header/components/Burger/index.js
@@ -1,0 +1,1 @@
+export { default } from './Burger.jsx';
