+++ src/modules/Header/components/UserMenu/index.js
@@ -1,0 +1,1 @@
+export { default } from './UserMenu.jsx';
