+++ src/modules/Price/Price.jsx
@@ -1,0 +1,5 @@
+const Price = () => {
+  return <div>Price</div>;
+};
+
+export default Price;
