+++ src/modules/Reviews/Reviews.jsx
@@ -1,0 +1,5 @@
+const Reviews = () => {
+  return <div>Reviews</div>;
+};
+
+export default Reviews;
